refactor(pieces): changed logic from calculating all possible moves to calculating all squares that are threatened
This is because taking into account all moves for pawns was preventing king from moving to squares below/above pawns even if those squares are not under threat.
This realization made me I could simplify a lot of the code and remove a lot of unnecesary filtering.
I also renamed get_possible_moves to get_attacking_squares to make it clearer what the purpose of this method is.
For Queens, Bishops, Rooks and Knights, legal moves are equal to their attacking squares but for pawns and kings this isn't the case.